[MOD] Reverse Proxy: Use Traefik instead of Nginx (#19)
After a recent deployment, with Nginx and Let's Encrypt companion
images being updated, I couldn't access anymore to any service.

Unable to troubleshoot the problem, I prefered to completely replace
Nginx by Traefik, which could be used in addition of Authelia in the
future, to provide MFA.

Another advantage of using Traefik is that its Docker image is
versioned, so less risk to get again troubles in a future deployment,
and being unable to rollback to a working state.